+++ encrypt.js
@@ -1,6 +1,8 @@
 var utils = require('./utils')
 
 module.exports = function (runtime, cb) {
+  utils.passthru(runtime, '[ -f env.sh ] || (touch env.sh && chmod 600 env.sh && git ignore env.sh)', { cwd: process.cwd() })
+
   if (!runtime.ghPagesEnv.GHPAGES_URL) {
     return cb(new Error('GHPAGES_URL was not set. Did you source env.sh first?'))
   }