+++ CHANGELOG.md
@@ -6,6 +6,15 @@
 - [ ] jRuby support https://spin.atomicobject.com/2013/04/24/bundler-jruby-complete/
 - [ ] Use orderoccurance in dev
 - [ ] Consider shipping all of Lanyon inside a Docker container
+- [ ] Utilize `passthru` more (vs `spawnSync`)
+- [ ] Add `travis` to Gemfile, and run it from our local shims
+
+# v0.0.15
+
+[Diff](https://github.com/kvz/lanyon/compare/v0.0.14...v0.0.15)
+
+- [x] Rename `vendor/bin` to `bin`
+- [x] Refactoring of deploy & encrypt
 
 # v0.0.14